Map View implemented
As described in issue #3, the leaflet mapping library has been imported
in the project in order to represent on a world-map, the position of
the results of a certain filter/search action. For that, following
steps have been done:
- leaflet-0.7.3 library included
- MarkerCluster leaflet plugin included.
- Button added in top bar to switch from/to  map/grid view.
- mapview.js has been created to host the code related to the map.
- in pinky.js, an initMap function has been added. Also, the code
snippet where the card UI elements of each article on the grid has been
encapsulated and named as _renderArticleCardHtml for reuse. Another,
and similar function, called _renderArticlePopupHtml has been created
to return the HTML content of the popups in the map. At the end of the
renderTags method, the function codeAddressFromArticle will be called
in order to geocode each of the results of the filter. The service used
for this was: http://www.mapquestapi.com/geocoding/